misc: Extract calculation of fees in a separate service
